Extract memberships.py from activations.py
Move ProcessedMemberships, MembershipBuilder, collect_memberships*,
and LM sampling helpers to spd/clustering/memberships.py.

activations.py retains: component_activations, ProcessedActivations,
process_activations, and dense filtering/coactivation helpers.
